% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linear regression function}
\arguments{
\item{formula}{formula is formula expression, like y ~ x1 + x2}
}
\value{
a linreg class
}
\description{
use RC class to compute the statistic
}
\details{
This is some description of this function.

you can use this function to calculate linear regression statistics,then predict the response.
}
\section{Fields}{

\describe{
\item{\code{formula}}{.}

\item{\code{data}}{.}

\item{\code{X}}{independent variables.}

\item{\code{y}}{dependent variable.}

\item{\code{beta_hat}}{Regressions coefficients.}

\item{\code{y_hat}}{The fitted values.}

\item{\code{e_hat}}{The residuals.}

\item{\code{df}}{The degrees of freedom.}

\item{\code{res_var}}{The residual variance.}

\item{\code{var_reg_coe}}{The variance of the regression coefficients.}

\item{\code{t_value}}{.}

\item{\code{p_value}}{.}
}}


